Move LeadMessage class into the model package
